# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EngTv IPTV API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List channels with optional filters
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
        - name: active_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get a channel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin: Auth ──────────────────────────────────────────────────────────
...
  /admin/channels:
    get:
      operationId: listAdminChannels
      tags: [channels, admin]
      summary: List all channels including inactive (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
        - name: source_id
          in: query
          schema:
            type: integer
        - name: language
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: is_healthy
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of all channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels, admin]
      summary: Create a channel (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteAllChannels
      tags: [channels, admin]
      summary: Delete ALL channels (admin, requires confirm header)
      security:
        - bearerAuth: []
      parameters:
        - name: x-confirm-delete-all
          in: header
          required: true
          schema:
            type: string
            enum: ["yes"]
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResult"
  /admin/channels/reorder:
    patch:
      operationId: reorderChannels
      tags: [channels, admin]
      summary: Reorder channels (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/channels/bulk-delete:
    post:
      operationId: bulkDeleteChannels
      tags: [channels, admin]
      summary: Delete multiple channels by ID (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIdsInput"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResult"
  /admin/channels/bulk-status:
    post:
      operationId: bulkUpdateChannelStatus
      tags: [channels, admin]
      summary: Enable or disable multiple channels by ID (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkStatusInput"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpdateResult"
  /admin/channels/bulk-category:
    post:
      operationId: bulkUpdateChannelCategory
      tags: [channels, admin]
      summary: Move multiple channels to a different category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCategoryMoveInput"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCategoryMoveResult"
...
  /admin/channels/{id}:
    patch:
      operationId: updateChannel
      tags: [channels, admin]
      summary: Update a channel (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteChannel
      tags: [channels, admin]
      summary: Delete a channel (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Admin: Sources ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ── App Settings ──────────────────────────────────────────────────────────
...
    Channel:
      type: object
      required: [id, name, stream_url, is_active, sort_order]
      properties:
        id:
          type: integer
        name:
          type: string
        stream_url:
          type: string
        logo_url:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        source_id:
          type: ["integer", "null"]
        external_id:
          type: ["string", "null"]
        is_active:
          type: boolean
        sort_order:
          type: integer
        language:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        created_at:
          type: string
        last_checked_at:
          type: ["string", "null"]
        is_healthy:
          type: ["boolean", "null"]
        health_error:
          type: ["string", "null"]
    ChannelInput:
      type: object
      required: [name, stream_url]
      properties:
        name:
          type: string
          minLength: 1
        stream_url:
          type: string
        logo_url:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        source_id:
          type: ["integer", "null"]
        external_id:
          type: ["string", "null"]
        is_active:
          type: boolean
        sort_order:
          type: ["integer", "null"]
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        stream_url:
          type: string
        logo_url:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        is_active:
          type: boolean
        sort_order:
          type: ["integer", "null"]

    # ── Bulk Operations ────────────────────────────────────────────────────────
    BulkIdsInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
    BulkStatusInput:
      type: object
      required: [ids, is_active]
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
        is_active:
          type: boolean
    BulkDeleteResult:
      type: object
      required: [deleted_count]
      properties:
        deleted_count:
          type: integer
    BulkUpdateResult:
      type: object
      required: [updated_count]
      properties:
        updated_count:
          type: integer
    BulkCategoryMoveInput:
      type: object
      required: [ids, category_id]
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
        category_id:
          type: ["integer", "null"]
    BulkCategoryMoveResult:
      type: object
      required: [updated_count]
      properties:
        updated_count:
          type: integer

    # ── Health Check ───────────────────────────────────────────────────────────
...
    ReorderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReorderItem"
    ReorderItem:
      type: object
      required: [id, sort_order]
      properties:
        id:
          type: integer
        sort_order:
          type: integer

    # ── Source ─────────────────────────────────────────────────────────────────
...